Mapping the attibutes and changing double to BigDecimal

+++ src/main/java/br/com/fiap/bank_project/entity/Bank.java
@@ -1,26 +1,27 @@
 package br.com.fiap.bank_project.entity;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
 
 public abstract class Bank implements Serializable {
 
-    protected double balance;
+    protected BigDecimal balance;
 
     protected Bank() {}
 
-    protected Bank(double balance) {
+    protected Bank(BigDecimal balance) {
         this.balance = balance;
     }
 
-    protected abstract void deposit(double value);
-    protected abstract void withdraw(double value);
+    protected abstract void deposit(BigDecimal value);
+    protected abstract void withdraw(BigDecimal value);
 
 
-    protected double getBalance() {
+    protected BigDecimal getBalance() {
         return balance;
     }
 
-    protected void setBalance(double balance) {
+    protected void setBalance(BigDecimal balance) {
         this.balance = balance;
     }
